requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albumentations=1.2.1=pypi_0
blas=1.0=mkl
blosc=1.21.0=h19a0ad4_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.07.19=haa95532_0
certifi=2022.6.15=py310haa95532_0
cffi=1.15.0=py310h2bbff1b_1
cfitsio=3.470=h2bbff1b_7
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py310h5588dad_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.1=py310h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pypi_0
cytoolz=0.11.0=py310h2bbff1b_0
dask-core=2022.5.0=py310haa95532_0
fonttools=4.34.4=pypi_0
freetype=2.10.4=hd328e21_0
fsspec=2022.3.0=py310haa95532_0
giflib=5.2.1=h62dcd97_0
git=2.34.1=haa95532_0
icc_rt=2019.0.0=h0cc432a_1
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.8.26=py310h6cc35ac_0
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
joblib=1.1.0=pypi_0
jpeg=9e=h2bbff1b_0
kiwisolver=1.4.4=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=h8ffe710_5
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
libzlib=1.2.12=h8ffe710_2
libzopfli=1.0.3=ha925a31_0
locket=1.0.0=py310haa95532_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib=3.5.2=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
networkx=2.7.1=pyhd3eb1b0_0
numpy=1.22.4=pypi_0
numpy-base=1.22.3=py310h206c741_0
opencv-python=4.6.0.66=pypi_0
opencv-python-headless=4.6.0.66=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.5=h8ffe710_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py310hf5e1058_0
partd=1.2.0=pyhd3eb1b0_1
pillow=9.0.1=py310hdc2b20a_0
pip=22.1.2=py310haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py310haa95532_0
python=3.10.5=hcf16a7b_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.20.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.12.0=py3.10_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pywavelets=1.3.0=py310h2bbff1b_0
pyyaml=6.0=py310h2bbff1b_0
qudida=0.0.4=pypi_0
requests=2.28.1=py310haa95532_0
scikit-image=0.19.2=py310h4ed8f06_0
scikit-learn=1.1.1=pypi_0
scipy=1.7.3=py310h6d2d95c_0
setuptools=63.2.0=py310h5588dad_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sqlite=3.39.2=h8ffe710_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.12=h8ffe710_0
toolz=0.11.2=pyhd3eb1b0_0
torchvision=0.13.0=py310_cu113
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=h191b570_0
urllib3=1.26.11=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
zfp=0.5.5=hd77b12b_6
zlib=1.2.12=h8cc25b3_2
zstd=1.4.9=h19a0ad4_0
